Tried adding binary search to FluxCapacitor
Disclaimer: I have somehow made it this far in my programming career without writing a single binary search, and I was never taught how to do so in school. So this could be horrifically wrong. It seems to be functioning (I get corrections, particularly when going up or down ramps), but from my pen and paper "tests" of the algorithm it looks like it may be running effectively the same operations twice if the size argument hits 3 (as size 2 will be next, and evaluates the same). So... that's a flaw.